% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordata.R
\name{cordata}
\alias{cordata}
\title{A Corr Data Function}
\usage{
cordata(NBA, pick_year)
}
\arguments{
\item{NBA}{Dataset. In this case is NBA stat.}

\item{pick_year}{Year of interest.}
}
\value{
A correlation of numberic values within NBA stat dataset.
}
\description{
A Corr Data Function
}
\examples{
cordata(NBA, 1995)
}
\keyword{Correlation}
