% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{the player for that year who scored the most points.
 @param year
 @export
 @example 
 f1(2011)}
\usage{
f1(x)
}
\description{
the player for that year who scored the most points.
 @param year
 @export
 @example 
 f1(2011)
}
