% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.R
\name{f2}
\alias{f2}
\title{the oldest player in the NBA data for that year.
 @param year
 @export
 @example 
 f2(2011)}
\usage{
f2(x)
}
\description{
the oldest player in the NBA data for that year.
 @param year
 @export
 @example 
 f2(2011)
}
