% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.R
\name{f3}
\alias{f3}
\title{the correlation matrix for all numeric variables in the data for that year.
 @param year
 @export
 @example 
 f3(2011)}
\usage{
f3(x)
}
\description{
the correlation matrix for all numeric variables in the data for that year.
 @param year
 @export
 @example 
 f3(2011)
}
