% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_numeric_correlation.R
\name{nba_numeric_correlation}
\alias{nba_numeric_correlation}
\title{An nbastat function}
\usage{
nba_numeric_correlation(year)
}
\arguments{
\item{year}{The year which will you wish to generate a correlation matrix for}
}
\description{
This function will generate a correlation matrix of numeric (continous) variables for a specific year.
}
\examples{
nba_numeric_correlation()

}
\keyword{correlation}
\keyword{nba}
\keyword{numeric}
