% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_old.R
\name{player_old}
\alias{player_old}
\title{An nbastat function}
\usage{
player_old(year)
}
\arguments{
\item{year}{The year for which you wish to determine the oldest player}
}
\description{
This function will return the oldest player in a specific year, if multiple players are of the same age, it will return the top row.
}
\examples{
player_old()
}
\keyword{old}
\keyword{oldest}
\keyword{player}
