% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_topscore.R
\name{player_topscore}
\alias{player_topscore}
\title{An nbastat function}
\usage{
player_topscore(year)
}
\arguments{
\item{year}{The year for which you wish to determine the top-scoring player}
}
\description{
This function will return the top-scoring player in a specific year.
}
\examples{
player_topscore()
}
\keyword{player}
\keyword{score}
\keyword{scoring}
\keyword{top}
