% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearoldest.R
\name{yearoldest}
\alias{yearoldest}
\title{An Oldest Function}
\usage{
yearoldest(NBA, pick_year)
}
\arguments{
\item{NBA}{Dataset. In this case is NBA stat.}

\item{pick_year}{Year of interest.}
}
\value{
The oldest player in NBA within the chosen year.
}
\description{
An Oldest Function
}
\examples{
function(NBA, 1995)
}
\keyword{oldest}
