% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearyoungest.R
\name{yearyoungest}
\alias{yearyoungest}
\title{A Youngest Function}
\usage{
yearyoungest(NBA, pick_year)
}
\arguments{
\item{NBA}{Dataset. In this case is NBA stat.}

\item{pick_year}{Year of interest.}
}
\value{
The youngest player in NBA within the chosen year.
}
\description{
A Youngest Function
}
\examples{
function(NBA, 1995)
}
\keyword{youngest}
